Remove useless info & restructure file
